Koala update
This update (as you can see in readme) isn't new. It for some reason was just never put onto github. The full files and updater has been running on the OFL servers for months and has been available to download from there.

The OFL updater that works with the OFL AC has to be downloaded from the OFL AC website.