Bundled updater and changed update locations to fix reported errors.
Upon community request, the updater has been fixed and updates will be downloaded from a legacy ABR Hosting path.
If you don't want to receive updates then you can use: `sm_updater 1` and updates will only be notified to you via the log file.

The updater used to be a seperate download hosted on the OFL website but since the OFL website has been taken offline along with the matchmaking servers and event services this meant it was traded around like a shiny pokemon card. Now it is available to all.